feat(suite): move slow Ble FW update to suite only desktop
